Optimized & Re structured ASC Mini App
