+++ lib/Navigator.dart
@@ -91,7 +91,8 @@ class AppNavigator extends StatelessWidget {
                 ),
               ),
               navigatorObservers: observers,
-              home: SKVHomeScreen(),
+              // home: SKVHomeScreen(),
+              home: ASCHomeScreen(),
               onGenerateRoute: (settings) {
                 final index = ["skvDetail", "hfdDetail"].indexOf(settings.name);
                 if (index > -1) {
